requirements.txt
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.4
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.7
install==1.3.5
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupytext==1.16.1
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
mpmath==1.3.0
multiprocess==0.70.16
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
p2j==1.3.2
packaging==24.0
pandas==2.2.1
parso==0.8.4
pathos==0.3.2
peewee==3.17.1
platformdirs==4.2.0
pox==0.3.4
ppft==1.7.6.8
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.5.1
pycparser==2.22
PyExecJS==1.5.1
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
Telethon==1.34.0
threadpoolctl==3.4.0
tokenizers==0.19.1
toml==0.10.2
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.0
translators==5.9.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13